Refactored application structure to resolve circular import issue
