+++ app/routes/__init__.py
@@ -1,1 +1,2 @@
-from app.routes.routes import routes
+# from app.routes.routes import routes
+from .routes import routes
